fix(docs): rename costEstimator module so it stops colliding by case
CostEstimator.tsx and costEstimator.ts sat in the same directory
differing only in case. On a case-insensitive filesystem (macOS, and
Windows) TypeScript resolves the import in Docs.tsx to the pure math
module rather than the component, so `tsc -b` fails:

  src/pages/Docs.tsx(5,10): error TS2305: Module './docs/CostEstimator'
  has no exported member 'CostEstimator'.

Linux CI is case-sensitive and resolves both specifiers correctly, so
this passed `make frontend-ci` while breaking the typecheck on every
macOS dev machine.

Renames the pure module to costMath.ts, which is what the code actually
is and no longer differs from the component only by case.
