+++ frontend/src/pages/docs/CostEstimator.tsx
@@ -7,7 +7,7 @@ import {
   formatTokens,
   formatUsd,
   type OperationCounts,
-} from './costEstimator'
+} from './costMath'
 
 const FIELD =
   'w-full bg-[#1c1c1c] border border-white/15 rounded px-3 py-2 text-sm text-white ' +
