+++ frontend/src/pages/docs/costMath.test.ts
@@ -7,7 +7,7 @@ import {
   formatTokens,
   formatUsd,
   type Operation,
-} from './costEstimator'
+} from './costMath'
 
 const OP: Operation[] = [
   {
