+++ frontend/src/pages/docs/costMath.ts
